LayoutTests/imported/w3c/web-platform-tests/mixed-content/gen/worker-module.http-rp/unset/worker-classic.https-expected.txt


FAIL Mixed-Content: Expects allowed for worker-classic to same-https origin and keep-scheme redirection from https context. assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
FAIL Mixed-Content: Expects allowed for worker-classic to same-https origin and no-redirect redirection from https context. assert_equals: The resource request should be 'allowed'. expected "allowed" but got "blocked"
PASS Mixed-Content: Expects blocked for worker-classic to same-http origin and keep-scheme redirection from https context.
PASS Mixed-Content: Expects blocked for worker-classic to same-http origin and no-redirect redirection from https context.
